corrected variable type

+++ Edge.h
@@ -5,18 +5,18 @@
 class Edge
 {
 private:
-    uint32_t idStart; // uint32_t since we expect a unique 32-bit unsigned number
+    uint32_t idStart;
     uint32_t idEnd;
-    float weight;
+    double weight;
 
 public:
     /* Constructor */
-    Edge(uint32_t idStart, uint32_t idEnd, float weight)
+    Edge(uint32_t idStart, uint32_t idEnd, double weight)
         : idStart(idStart), idEnd(idEnd), weight(weight) {}
     ~Edge() = default;
 
     /* Getters */
-    float getWeight() const { return weight; }
+    double getWeight() const { return weight; }
     uint32_t getStartId() const { return idStart; }
     uint32_t getEndId() const { return idEnd; }
 };
